+++ packages/core/lib/src/application/firebase/firebase_event.dart
@@ -1,0 +1,6 @@
+part of 'firebase_bloc.dart';
+
+@freezed
+abstract class FirebaseEvent with _$FirebaseEvent {
+  const factory FirebaseEvent.onChangeColor(AppBarTitle color) = OnChangeColor;
+}