+++ packages/core/lib/src/application/firebase/firebase_state.dart
@@ -1,0 +1,12 @@
+part of 'firebase_bloc.dart';
+
+@freezed
+abstract class FirebaseState with _$FirebaseState {
+  const factory FirebaseState({
+    required String color,
+  }) = _FirebaseState;
+
+  factory FirebaseState.initial() => FirebaseState(
+    color: '',
+  );
+}