+++ packages/models/lib/src/enums/export.dart
@@ -1,4 +1,5 @@
 export 'feedback/feedback_field.dart';
+export 'firebase/firebase_config_keys.dart';
 export 'menu/menu_items.dart';
 export 'product_tour/product_tour_step.dart';
 export 'storage/storage_keys.dart';